Refactor UI library components to utilize shared data files and improve clipboard functionality
