+++ src/pages/UiLibraries/Chakraui/ChakrauiInstallation.jsx
@@ -3,40 +3,13 @@ import { useState } from "react";
 import { SiChakraui } from "react-icons/si";
 import { HiCheckCircle } from "react-icons/hi";
 import { IoCopyOutline } from "react-icons/io5";
-
-const installSteps = [
-    {
-        title: "1. Install Chakra UI",
-        command: "npm install @chakra-ui/react @emotion/react @emotion/styled framer-motion",
-        description: "Install Chakra UI along with Emotion and Framer Motion (required peer dependencies).",
-    },
-    {
-        title: "2. Set up ChakraProvider",
-        command: `// main.jsx or index.js
-import * as React from "react";
-import * as ReactDOM from "react-dom/client";
-import { ChakraProvider } from "@chakra-ui/react";
-import App from "./App";
-
-const root = ReactDOM.createRoot(document.getElementById("root"));
-root.render(
-  <ChakraProvider>
-    <App />
-  </ChakraProvider>
-);`,
-        description: "Wrap your app with ChakraProvider to enable theming and styles.",
-        isCodeBlock: true,
-    },
-];
+import { handleCopy } from "@/components/shared/CopyToClipboard";
+import { packageManagers, installSteps } from "@/pages/UiLibraries/Chakraui/ChakrauiData";
+import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 
 export const ChakrauiInstallation = () => {
     const [copied, setCopied] = useState(null);
-
-    const handleCopy = async (text, idx) => {
-        await navigator.clipboard.writeText(text);
-        setCopied(idx);
-        setTimeout(() => setCopied(null), 1500);
-    };
+    const [selectedTabs, setSelectedTabs] = useState({});
 
     return (
         <div className="space-y-8 p-4 md:p-8">
@@ -67,8 +40,15 @@ export const ChakrauiInstallation = () => {
                             </div>
                             <div className="relative">
                                 <button
-                                    onClick={() => handleCopy(step.command, idx)}
-                                    className="text-muted-foreground hover:text-primary transition"
+                                    onClick={() => handleCopy(
+                                        typeof step.command === "string"
+                                            ? step.command
+                                            : step.command[selectedTabs[idx] || "npm"], // pick per-step tab
+                                        setCopied,
+                                        idx
+                                    )}
+                                    className="text-muted-foreground hover:text-[#319795] transition"
+                                    title="Copy to clipboard"
                                 >
                                     {copied === idx ? (
                                         <HiCheckCircle className="w-6 h-6 text-green-600" />
@@ -84,9 +64,36 @@ export const ChakrauiInstallation = () => {
                             </div>
                         </div>
 
-                        <div className="bg-muted/50 border border-[#cccccc] rounded-md p-3 mt-2 font-mono text-sm whitespace-pre-wrap overflow-x-auto">
+                        {/* <div className="bg-muted/50 border border-[#cccccc] rounded-md p-3 mt-2 font-mono text-sm whitespace-pre-wrap overflow-x-auto">
                             {step.command}
-                        </div>
+                        </div> */}
+                        {/* Tabs */}
+                        {step.all ? (
+                            <pre className="bg-muted/50 border border-gray-300 rounded-md p-3 font-mono text-sm whitespace-pre-wrap overflow-x-auto">
+                                {step.command}
+                            </pre>
+                        ) : (
+                            <Tabs value={selectedTabs[idx] || "npm"} onValueChange={(value) => setSelectedTabs((prev) => ({ ...prev, [idx]: value }))} className="w-full">
+                                <TabsList className="">
+                                    {packageManagers.map((pm) => (
+                                        <TabsTrigger key={pm.key} value={pm.key}>
+                                            <span className="flex items-center gap-1 text-xs cursor-pointer">
+                                                {pm.icon}
+                                                {pm.label}
+                                            </span>
+                                        </TabsTrigger>
+                                    ))}
+                                </TabsList>
+
+                                {packageManagers.map((pm) => (
+                                    <TabsContent key={pm.key} value={pm.key}>
+                                        <pre className="bg-muted/50 border border-gray-300 rounded-md p-3 font-mono text-sm whitespace-pre-wrap overflow-x-auto">
+                                            {step.command[pm.key]}
+                                        </pre>
+                                    </TabsContent>
+                                ))}
+                            </Tabs>
+                        )}
                     </div>
                 ))}
             </div>
